% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_detection.R
\name{consensus_detection}
\alias{consensus_detection}
\title{Remove ambiguous detections}
\usage{
consensus_detection(
  detection,
  by = "overlap",
  filter = "max",
  cores = 1,
  pb = TRUE
)
}
\arguments{
\item{detection}{Data frame or selection table (using the warbleR package's format, see \code{\link[warbleR]{selection_table}}) with the output of \code{\link{label_detection}} containing the start and end of the signals. Must contain at least the following columns: "sound.files", "selec", "start", "end" and "detection.class" (the last one is generated by \code{\link{label_detection}}). It must also contain the column indicated in the 'by' argument (which is 'overlap' by default).}

\item{by}{Character vector of length 1 indicating a column in 'detection' that will be used to filter detections. Must refer to a numeric column. Default is 'overlap', which is returned by \code{\link{label_detection}}.}

\item{filter}{Character vector of length 1 indicating the criterion used to filter the column referred to by the 'by' argument. Current options are 'max' (maximum) and 'min' (minimum). Default is 'max'.}

\item{cores}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}
}
\value{
A data frame or selection table (if 'detection' was also a selection table, warbleR package's format, see \code{\link[warbleR]{selection_table}}) as in 'detection' but removing ambiguous detections (split and merged positives).
}
\description{
\code{consensus_detection} removes ambiguous detections
}
\details{
This function removes ambiguous detections keeping only the one that maximizes a criterion given by 'filter'. By default it keeps the detection with the highest overlap to the reference signal. It works on the output of \code{\link{label_detection}}. Mostly useful when several detections match the same reference as in the case of template detection with multiple templates (see \code{\link{template_detector}}). Keep in mind that the argument 'solve.ambiguous' must be \code{FALSE} to keep those ambiguous detections.
}
\examples{
{
  # load example data
  data("lbh1", "lbh_reference")

  # save sound files
  tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))

  # template for the first sound file in 'lbh_reference'
  templ1 <- lbh_reference[lbh_reference$sound.files == "lbh1.wav" & lbh_reference$selec == 11, ]

  # generate template correlations
  tc <- template_correlator(
    templates = templ1, path = tempdir(),
    files = "lbh1.wav"
  )

  # template detection
  td <- template_detector(template.correlations = tc, threshold = 0.35)

  # this detection generates 2 split positives
  diagnose_detection(
    reference = lbh_reference[lbh_reference$sound.files == "lbh1.wav", ],
    detection = td,
    solve.ambiguous = FALSE
  )

  # label detection
    ltd <- label_detection(
    reference = lbh_reference[lbh_reference$sound.files == "lbh1.wav", ],
    detection = td, 
    solve.ambiguous = FALSE
    )

  # now they can be filtered to keep the detection with the highest score for each split
  ftd <- consensus_detection(ltd, by = "scores")

  # splits must be 0
  diagnose_detection(
    reference = lbh_reference[lbh_reference$sound.files == "lbh1.wav", ],
    detection = ftd,
    solve.ambiguous = FALSE
  )
}

}
\references{
Araya-Salas, M., Smith-Vidaurre, G., Chaverri, G., Brenes, J. C., Chirino, F., Elizondo-Calvo, J., & Rico-Guevara, A. (2023). ohun: An R package for diagnosing and optimizing automatic sound event detection. Methods in Ecology and Evolution, 14, 2259–2271. https://doi.org/10.1111/2041-210X.14170
}
\seealso{
\code{\link{label_detection}}, \code{\link{template_detector}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr}).
}
