% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envelopes.R
\name{envelopes}
\alias{envelopes}
\title{Class 'envelopes': list of absolute amplitude envelopes}
\description{
Class for absolute amplitude envelopes
}
\details{
An object of class \code{envelopes} created by \code{\link{get_envelopes}} is a list with sound files absolute amplitude envelopes and metadata
}
\seealso{
\code{\link{get_envelopes}}
}
