% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_reference.R
\name{feature_reference}
\alias{feature_reference}
\title{Extract quantitative features of references}
\usage{
feature_reference(reference, path = NULL, by.sound.file = FALSE)
}
\arguments{
\item{reference}{Data frame or 'selection.table' (following the warbleR package format) with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end". If frequency range columns are included ("bottom.freq" and "top.freq") these are also used to characterize reference selections.}

\item{path}{Character string containing the directory path where the sound files are located. If supplied then duty cycle features are returned.}

\item{by.sound.file}{Logical argument to control whether features are summarized across sound files (when \code{by.sound.file = FALSE}, and more than 1 sound file is included in 'reference') or shown separated by sound file. Default is \code{FALSE}.}
}
\value{
The function returns the mean, minimum and maximum duration of selections and gaps (time intervals between selections). If frequency range columns are included in the reference table (i.e. "bottom.freq" and "top.freq") the minimum bottom frequency ('min.bottom.freq') and the maximum top frequency ('max.top.freq') are also estimated. Finally, if the path to the sound files in 'reference' is supplied the duty cycle the mean, minimum and maximum. If `by.sound.file = FALSE` a matrix with features in rows is returned. Otherwise a data frame is returned in which each row correspond to a sound file.
}
\description{
\code{feature_reference} extracts quantitative characteristics of a reference table
}
\details{
The function extract quantitative features from reference tables that can inform the range of values to be used in a energy-based detection optimization routine. Features related to selection duration can be used to set the 'max.duration' and 'min.duration' values, frequency related features can inform banpass values, gap related features inform hold time values and duty cycle can be used to evaluate performance. Note that duty cycle is not included as a diagnostic in other functions but it can also be estimated on detections using `feature_reference`.
}
\examples{
{
# load data
data("lbh_reference")

# summary across sound files
feature_reference(reference = lbh_reference)

# summary across sound files
feature_reference(reference = lbh_reference, by.sound.file = TRUE)
}
}
\references{
{
Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.
}
}
\seealso{
\code{\link{optimize_energy_detector}}, \code{\link{optimize_template_detector}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
