% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbh2-data.R
\docType{data}
\name{lbh2}
\alias{lbh2}
\title{Long-billed hermit recording}
\format{
An object of class \code{Wave} of length 110250.
}
\source{
Marcelo Araya-Salas
}
\usage{
data(lbh2)
}
\description{
\code{lbh2}  a wave object with long-billed hermit (\emph{Phaethornis longitrostris}) songs extracted from xeno-canto's '154129' recording.
}
\keyword{datasets}
