% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbh_1-data.R
\docType{data}
\name{lbh_1}
\alias{lbh_1}
\title{Long-billed hermit recording}
\format{
An object of class \code{Wave} of length 110250.
}
\source{
Marcelo Araya Salas, warbleR
}
\usage{
data(lbh_2)
}
\description{
\code{lbh_1} a wave object with is a long-billed hermit songs extracted from xeno-cantos '154129' recording.
}
\keyword{datasets}
