% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbh_2-data.R
\docType{data}
\name{lbh_2}
\alias{lbh_2}
\title{Long-billed hermit recording}
\format{
An object of class \code{Wave} of length 110250.
}
\source{
Marcelo Araya Salas, warbleR
}
\usage{
data(lbh_2)
}
\description{
\code{lbh_2} a wave object with is a long-billed hermit songs extracted from xeno-cantos '154138' recording.
}
\keyword{datasets}
