% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_envelopes.R
\name{print.envelopes}
\alias{print.envelopes}
\title{print method for class \code{envelopes}}
\usage{
\method{print}{envelopes}(x, ...)
}
\arguments{
\item{x}{Object of class \code{envelopes}, generated by \code{\link{get_envelopes}}.}

\item{...}{further arguments passed to or from other methods. Ignored when printing envelopes.}
}
\value{
Prints a summary of an object of class 'envelopes'.
}
\description{
print method for class \code{envelopes}
}
\keyword{internal}
