% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{print.template_correlations}
\alias{print.template_correlations}
\title{print method for class \code{template_correlations}}
\usage{
\method{print}{template_correlations}(x, ...)
}
\arguments{
\item{x}{Object of class \code{template_correlations}, generated by \code{\link{template_correlator}}.}

\item{...}{further arguments passed to or from other methods. Ignored when printing 'template_correlations' class objects.}
}
\value{
Prints a summary of an object of class 'template_correlations'.
}
\description{
print method for class \code{template_correlations}
}
\keyword{internal}
