% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_detection.R
\name{diagnose_detection}
\alias{diagnose_detection}
\title{Evaluate the performance of a signal detection procedure}
\usage{
diagnose_detection(reference, detection, by.sound.file = FALSE,
time.diagnostics = FALSE)
}
\arguments{
\item{reference}{Data frame or 'selection.table' with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".}

\item{detection}{Data frame or 'selection.table' with the detections (start and end of the signals) that will be compared against the 'reference' selections. Must contained at least the following columns: "sound.files", "selec", "start" and "end".}

\item{by.sound.file}{Logical argument to control whether performance diagnostics are summarized across sound files (when \code{by.sound.file = FALSE}, when more than 1 sound file is included in 'reference') or shown separated by sound file. Default is \code{FALSE}.}

\item{time.diagnostics}{Logical argument to control if diagnostics related to the duration of the signals ("mean.duration.true.positives", "mean.duration.false.positives", "mean.duration.false.negatives" and "proportional.duration.true.positives") are returned (if \code{TRUE}). Default is \code{FALSE}.}
}
\value{
A data frame including the following detection performance diagnostics:
\itemize{
 \item \code{true.positives}: number of detections that correspond to signals referenced in 'reference'. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.
 \item \code{false.positives}: number of detections that don't match any of the signals referenced in 'reference'. In a perfect detection routine it should be 0.
 \item \code{false.negatives}: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0.
 \item \code{split.positives}: number of signals referenced in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.
 \item \code{mean.duration.true.positives}: mean duration of true positives (in s). Only included when \code{time.diagnostics = TRUE}.
 \item \code{mean.duration.false.positives}: mean duration of false positives (in s). Only included when \code{time.diagnostics = TRUE}.
 \item \code{mean.duration.false.negatives}: mean duration of false negatives (in s). Only included when \code{time.diagnostics = TRUE}.
 \item \code{proportional.duration.true.positives}: ratio of total duration of true positives to the total duration of signals referenced in 'reference'. In a perfect detection routine it should be 1. Only included when \code{time.diagnostics = TRUE}.
 \item \code{sensitivity}: Proportion of signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.
 \item \code{specificity}: Proportion of detections that correspond to signals referenced in 'reference' that were detected. In a perfect detection routine it should be 1.
 }
}
\description{
\code{diagnose_detection} evaluates the performance of a signal detection procedure comparing the output selection table to a reference selection table
}
\details{
The function evaluates the performance of a signal detection procedure by comparing its output selection table to a reference selection table in which all signals of interest have been selected.
}
\examples{
\dontrun{
# Save example files into temporary working directory
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# run detection
detec <- energy_detector(path = tempdir(), threshold = 6, ssmooth = 300,
bp = c(2, 9), wl = 300, min.duration = 0.09)

# diagnose detection
diagnose_detection(reference = lbh_selec_reference,
detection = detec)

# include time diagnostics
diagnose_detection(reference = lbh_selec_reference,
detection = detec, time.diagnostics = TRUE)

# by sound file
diagnose_detection(reference = lbh_selec_reference,
detection = detec, by.sound.file = TRUE)
}
}
\references{
{
Araya-Salas, M. (2020), ohun: automatic detection of acoustic signals. R package version 0.1.0.
}
}
\seealso{
\code{\link{optimize_auto_detec}}, \code{\link{optimize_find_peaks}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
