% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohun-package.R
\docType{package}
\name{ohun}
\alias{ohun}
\title{ohun: Automatic detection of acoustic signals}
\description{
ohun is intended to facilitate the automatic detection of acoustic signals, providing functions to diagnose and optimize detection routines. Detections from other software can also be explored and optimized.
}
\details{
The main features of the package are:
  \itemize{
  \item The batch processing of sound files for signal detection  
  \item The usage of annotations for detection optimization and diagnostic
  }
  
The package offers functions to:
  \itemize{
  \item Energy-based detection
  \item Template-based detection
  \item Diagnose detection precision
  \item Optimize detection routines based on reference annotations
  }
  
Most of the functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency.

License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas & Grace Smith Vidaurre
  
  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
