% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohun-package.R
\docType{package}
\name{ohun}
\alias{ohun-package}
\alias{ohun}
\title{ohun: Optimizing sound event detection}
\description{
ohun is intended to facilitate the automated detection of sound events, providing functions to diagnose and optimize detection routines. Detections from other software can also be explored and optimized.
}
\details{
The main features of the package are:
  \itemize{
  \item The use of reference annotations for detection optimization and diagnostic
  \item The use of signal detection theory diagnostic parameters to evaluate detection performance
  \item The batch processing of sound files for improve computational performance
  }

The package offers functions for:
  \itemize{
  \item Energy-based detection
  \item Template-based detection
  \item Diagnose detection precision
  \item Improve detection by adjusting parameters to optimize accuracy
  }

All functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency. The package works on sound files in '.wav', '.mp3', '.flac' and '.wac' format.

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/ohun/}
  \item \url{https://github.com/ropensci/ohun/}
  \item Report bugs at \url{https://github.com/ropensci/ohun/issues/}
}

}
\author{
Marcelo Araya-Salas

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\keyword{internal}
