% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_template_detector.R
\name{optimize_template_detector}
\alias{optimize_template_detector}
\title{Optimize acoustic template detection}
\usage{
optimize_template_detector(template.correlations, reference, threshold,
parallel = 1, pb = FALSE, by.sound.file = FALSE, previous.output = NULL)
}
\arguments{
\item{template.correlations}{An object of class 'template_correlations' (generated by \code{\link{template_correlator}}) in which to optimize detections. Must contain data for all sound files as in 'reference'. It can also contain data for additional sound files. In this case the routine assumes that no signals are found in those files, so detection from those files are all false positives.}

\item{reference}{Data frame or 'selection.table' (following the warbleR package format) with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end". \strong{It must contain the reference selections that will be used for detection optimization}.}

\item{threshold}{Numeric vector of length > 1 with values between 0 and 1 specifying the correlation threshold for detecting signal occurrences (i.e. correlation peaks). Must be supplied. \strong{Several values should be supplied for optimization}.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar and messages. Default is \code{TRUE}.}

\item{by.sound.file}{Logical to control if diagnostics are calculated for each sound file independently (\code{TRUE}) or for all sound files combined (\code{FALSE}, default).}

\item{previous.output}{Data frame with the output of a previous run of this function. This will be used to include previous results in the new output and avoid recalculating detection performance for parameter combinations previously evaluated.}
}
\value{
A data frame in which each row shows the result of a detection job for each cutoff value, including the following diagnostic metrics:
\itemize{
 \item \code{true.positives}: number of signals in 'reference' that correspond to any detection. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.
 \item \code{false.positives}: number of detections that don't match any of the signals in 'reference'. In a perfect detection routine it should be 0.
 \item \code{false.negatives}: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0.
 \item \code{split.positives}: number of signals in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.
 \item \code{merged.positives}: number of signals in 'detection' that were overlapped by more than 1 detection (i.e. signals that were merged). In a perfect detection routine it should be 0.
 \item \code{sensitivity}: Proportion of signals in 'reference' that were detected. In a perfect detection routine it should be 1.
 \item \code{specificity}: Proportion of detections that correspond to signals in 'reference' that were detected. In a perfect detection routine it should be 1.
 }
}
\description{
\code{\link{optimize_template_detector}} optimizes acoustic template detection
}
\details{
This function takes a selections data frame or 'selection_table' ('X') and the output of a \code{\link{auto_detec}} routine ('xc.output') and estimates the detection performance for different detection parameter combinations. This is done by comparing the position in time of the detection to those of the reference selections in 'X'. The function returns several diagnostic metrics to allow user to determine which parameter values provide a detection that more closely matches the selections in 'X'. Those parameters can be later used for performing a more efficient detection using \code{\link{optimize_template_detector}}.
}
\examples{
{
# Save sound files to temporary working directory
data(list = c("Phae.long4", "Phae.long2", "Phae.long3", "lbh_selec_reference"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))

# template for the fourth sound file in 'lbh_selec_reference'
templ <- lbh_selec_reference[11, ]

# generate template correlations
tc <- template_correlator(templates = templ, path = tempdir(),
files = "Phae.long4.wav")

# using 2 threshold
optimize_template_detector(template.correlations = tc, reference =
lbh_selec_reference[lbh_selec_reference$sound.files == "Phae.long4.wav", ],
threshold = c(0.2, 0.5))

# using several thresholds
optimize_template_detector(template.correlations = tc,
reference = lbh_selec_reference[lbh_selec_reference$sound.files == "Phae.long4.wav", ],
 threshold = seq(0.5, 0.9, by = 0.05))

 # template for the second and third sound file in 'lbh_selec_reference'
 templ <- lbh_selec_reference[c(4, 7), ]

 # generate template correlations
 tc <- template_correlator(templates = templ, path = tempdir(),
 files = c("Phae.long2.wav", "Phae.long3.wav"))

optimize_template_detector(template.correlations = tc, reference =
  lbh_selec_reference[lbh_selec_reference$sound.files \%in\% c("Phae.long2.wav", "Phae.long3.wav"), ],
 threshold = seq(0.5, 0.7, by = 0.1))

 # showing diagnostics by sound file
 optimize_template_detector(template.correlations = tc, reference =
 lbh_selec_reference[lbh_selec_reference$sound.files \%in\% c("Phae.long2.wav", "Phae.long3.wav"), ],
 threshold = seq(0.5, 0.7, by = 0.1), by.sound.file = TRUE)
}

}
\references{
{
Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.
}
}
\seealso{
\code{\link{optimize_auto_detec}}, \code{\link{find_peaks}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr}).
}
