% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_acoustic_data.R
\name{summarize_acoustic_data}
\alias{summarize_acoustic_data}
\title{alternative name for \code{\link{feature_acoustic_data}}}
\usage{
summarize_acoustic_data(path = ".", digits = 2)
}
\description{
alternative name for \code{\link{feature_acoustic_data}}
}
\details{
see \code{\link{summarize_acoustic_data}} for documentation. \code{\link{feature_acoustic_data}} will be deprecated in future versions.
}
\keyword{internal}
