% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_reference.R
\name{summarize_reference}
\alias{summarize_reference}
\title{alternative name for \code{\link{feature_reference}}}
\usage{
summarize_reference(
  reference,
  path = NULL,
  by.sound.file = FALSE,
  units = c("ms", "kHz"),
  digits = 2
)
}
\description{
alternative name for \code{\link{feature_reference}}
}
\details{
see \code{\link{summarize_reference}} for documentation. \code{\link{feature_reference}} will be deprecated in future versions.
}
\keyword{internal}
