% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_detector.R
\name{template_detector}
\alias{template_detector}
\title{Acoustic template detection from time-frequency cross-correlations}
\usage{
template_detector(template.correlations, parallel = 1, threshold, pb = TRUE,
 verbose = TRUE)
}
\arguments{
\item{template.correlations}{object of class 'template_correlations' generated by \code{\link{template_correlator}} containing the correlation score vectors.}

\item{parallel}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{threshold}{Numeric vector of length 1 with a value between 0 and 1 specifying the correlation threshold for detecting signal occurrences (i.e. correlation peaks). Must be supplied. Correlation scores are forced to between 0 and 1 (by converting negative scores to 0). 0 and 1 represent the lowest and highest similarity to the template respectively.}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}

\item{verbose}{Logical argument to control if some summary messages are printed to the console.}
}
\value{
The function returns a data frame with the start and end and correlation score for the
detected signals.
}
\description{
\code{template_detector} find signal occurrences in cross-correlation vectors from \code{\link{template_correlator}}
}
\details{
This function infers signals occurrences from cross-correlation scores along sound files. Correlation scores must be generated first using \code{\link{template_correlator}}.
}
\examples{
{
# load example data
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "lbh_selec_table2", "comp_matrix"))

# save sound files
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))

# template for the first sound file in 'lbh_selec_reference'
templ1 <- lbh_selec_reference[1, ]

# generate template correlations
tc <- template_correlator(templates = templ1, path = tempdir(), files = "Phae.long1.wav")

# template detection
td <- template_detector(template.correlations = tc, threshold = 0.4)

# diagnose detection
diagnose_detection(reference =
lbh_selec_reference[lbh_selec_reference$sound.files == "Phae.long1.wav", ],
detection = td)

# template for the second and third sound file in 'lbh_selec_reference'
# which have similar song types
templ2 <- lbh_selec_reference[4, ]

# generate template correlations
tc <- template_correlator(templates = templ2, path = tempdir(),
files = c("Phae.long1.wav", "Phae.long2.wav"))

# template detection
td <- template_detector(template.correlations = tc, threshold = 0.3)

# diagnose detection
diagnose_detection(reference =
lbh_selec_reference[lbh_selec_reference$sound.files \%in\% c("Phae.long3.wav",
"Phae.long2.wav"), ],
detection = td)
}
}
\references{
{
Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.
}
}
\seealso{
\code{\link{energy_detector}}, \code{\link{template_correlator}}, \code{\link{optimize_template_detector}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
